Move retention-related tests to retention_test.go (#414)
